EL
Weapon System Works
Hit Effect Works
Ammo count/Reload fix
Death presentation fix

+++ RedTeamGame/Assets/Scripts/gameManager.cs
@@ -178,7 +178,7 @@ public class gameManager : MonoBehaviour
     }
     public void UpdateShiieldUi()
     {
-        float tmp = playerScript.shieldAmount / playerScript.shieldAmountOrg;
+        float tmp = (float)playerScript.shieldAmount / playerScript.shieldAmountOrg;
         LShield.fillAmount = tmp;
         RShield.fillAmount = tmp;
     }
